Update to latest Kotlinx datetime and move to stdlib Instant and Clock.

+++ core/src/commonMain/kotlin/maryk/core/properties/types/Date.kt
@@ -5,7 +5,7 @@ import maryk.core.extensions.bytes.initInt
 import maryk.core.extensions.bytes.writeBytes
 
 internal fun LocalDate.writeBytes(writer: (byte: Byte) -> Unit) {
-    this.toEpochDays().writeBytes(writer)
+    this.toEpochDays().toInt().writeBytes(writer)
 }
 
 /** Reads a date from bytes [reader] */
